First test of cdev branch
This is the first in hopefully a long series of commits that will attemt
to convert the majority of GWar code to c++/SDL. It will be a good
excercise and hopefully provide additional benefit to the game itself.
Most game logic code will be reused, but the structure may be
reorganized.